backend/requirements_temp.txt
aiofiles==24.1.0
aiosqlite==0.20.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==3.7.1
audioop-lts==0.2.2
blinker==1.9.0
Brotli==1.1.0
CacheControl==0.14.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudevents==1.12.0
coverage==7.12.0
cryptography==46.0.5
deprecation==2.1.0
fastapi==0.115.0
ffmpy==0.6.4
filelock==3.20.0
firebase_admin==7.3.0
firebase_functions==0.5.0
Flask==3.1.2
flask-cors==6.0.2
fsspec==2025.10.0
functions-framework==3.10.1
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-firestore==2.25.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-events==0.5.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.1.1
groovy==0.1.2
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==25.1.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numpy==2.3.4
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.12.1
PyMySQL==1.1.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rich==14.2.0
ruff==0.14.3
safehttpx==0.1.7
safetensors==0.6.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
sympy==1.14.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.42.0
uvicorn-worker==0.4.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
